:truck: move components out of pages
