+++ src/pages/contact.jsx
@@ -1,39 +1,17 @@
 import React from 'react'
 import ContentContainer from '../components/ContentContainer'
 import BioCard from '../components/BioCard'
-import styled from 'styled-components'
-import tw from 'tailwind.macro'
 import { BtnOutline, Btn } from '../components/button'
 import Panel from '../components/Panel'
-import { rhythm } from '../../config/typography'
 import PageHeading from '../components/PageHeading'
-import { Subtitle } from '.'
-import SEO from '../components/SEO'
-
-const TextInput = styled.input`
-  ${tw`block p-2 text-white bg-gray-300 border-2 border-gray-300 rounded-lg w-72`};
-  :hover,
-  :focus {
-    ${tw`bg-gray-600 border-primary`}
-  }
-`
-const Label = styled.label`
-  ${tw`block my-2`}
-`
-export const DualWrapper = styled.div`
-  margin-top: ${rhythm(1)};
-  margin-bottom: ${rhythm(1)};
-  @media screen and (min-width: ${props => props.theme.breakpoints.tablet}) {
-    * + * {
-      margin: ${rhythm(2 / 3)};
-    }
-  }
-`
+import { Subtitle } from '../components/Hero'
+import { Label, TextInput, DualWrapper } from '../components/TextInput'
+import Seo from '../components/Seo'
 
 const Contact = () => {
   return (
     <>
-      <SEO
+      <Seo
         title={'Contact'}
         keywords={[`blog`, `gatsby`, `javascript`, `react`]}
       />
@@ -62,7 +40,7 @@ const Contact = () => {
                 placeholder="name@example.com"
               />
             </Label>
-            <label>
+            <Label>
               Subject
               <TextInput
                 type="text"
@@ -70,7 +48,7 @@ const Contact = () => {
                 id="subject"
                 placeholder="Subject"
               />
-            </label>
+            </Label>
             <Label>
               Message
               <TextInput
